9a_LangGraph/pyproject.toml
[project]
name = "10-langgraph"
version = "0.1.0"
description = ""
package-mode = false
authors = [
    {name = "Danish Mustafa",email = "danishjajja86@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langgraph (>=0.4.8,<0.5.0)",
    "typing-extensions (>=4.14.0,<5.0.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-core (>=0.3.68,<0.4.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pypdf (>=5.7.0,<6.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"